Fix up redundant files
