+++ package.json
@@ -15,7 +15,7 @@
     "test": "yarn evm:test && yarn ciphernode:test",
     "test:integration": "./tests/basic_integration/test.sh",
     "coverage": "yarn evm:coverage",
-    "ciphernode:launch": "cd packages/ciphernode && ./scripts/launch.sh",
+    "enclave": "cd packages/ciphernode && ./scripts/launch.sh",
     "ciphernode:lint": "cd packages/ciphernode && cargo fmt -- --check",
     "ciphernode:add": "cd packages/evm && yarn ciphernode:add",
     "ciphernode:remove": "cd packages/evm && yarn ciphernode:remove",
