--- packages/ciphernode/enclave/src/lib.rs
@@ -1,24 +1,0 @@
-pub mod commands;
-pub mod cli;
-
-const OWO: &str = r#"
-      ___           ___           ___                         ___                         ___     
-     /\__\         /\  \         /\__\                       /\  \          ___          /\__\    
-    /:/ _/_        \:\  \       /:/  /                      /::\  \        /\  \        /:/ _/_   
-   /:/ /\__\        \:\  \     /:/  /                      /:/\:\  \       \:\  \      /:/ /\__\  
-  /:/ /:/ _/_   _____\:\  \   /:/  /  ___   ___     ___   /:/ /::\  \       \:\  \    /:/ /:/ _/_ 
- /:/_/:/ /\__\ /::::::::\__\ /:/__/  /\__\ /\  \   /\__\ /:/_/:/\:\__\  ___  \:\__\  /:/_/:/ /\__\
- \:\/:/ /:/  / \:\~~\~~\/__/ \:\  \ /:/  / \:\  \ /:/  / \:\/:/  \/__/ /\  \ |:|  |  \:\/:/ /:/  /
-  \::/_/:/  /   \:\  \        \:\  /:/  /   \:\  /:/  /   \::/__/      \:\  \|:|  |   \::/_/:/  / 
-   \:\/:/  /     \:\  \        \:\/:/  /     \:\/:/  /     \:\  \       \:\__|:|__|    \:\/:/  /  
-    \::/  /       \:\__\        \::/  /       \::/  /       \:\__\       \::::/__/      \::/  /   
-     \/__/         \/__/         \/__/         \/__/         \/__/        ~~~~           \/__/    
-                                                                      
-"#;
-
-pub fn owo() {
-    println!("\n\n\n\n\n{}", OWO);
-    println!("\n\n\n\n");
-}
-
-
