--- packages/ciphernode/scripts/aggregator.sh
@@ -1,3 +1,0 @@
-#!/usr/bin/env bash
-
-RUSTFLAGS="-A warnings" cargo run --bin aggregator -- "$@"
